application-backend-server/requirements.txt
flask
requests
python-jose